[REFACTOR] Remove wrapper from utility pages
